refactor: move conversion functions until separate file
